wrote tests for existing action creators
